refactor: move DescriptorNotFoundException to common
