+++ dagger-core/src/main/java/io/odpf/dagger/core/processors/common/DescriptorManager.java
@@ -1,7 +1,7 @@
 package io.odpf.dagger.core.processors.common;
 
 import io.odpf.dagger.common.core.StencilClientOrchestrator;
-import io.odpf.dagger.core.exception.DescriptorNotFoundException;
+import io.odpf.dagger.common.exceptions.DescriptorNotFoundException;
 import com.gojek.de.stencil.client.StencilClient;
 import com.google.protobuf.Descriptors;
