+++ dagger-core/src/main/java/io/odpf/dagger/core/processors/external/AsyncConnector.java
@@ -11,7 +11,7 @@ import org.apache.flink.types.Row;
 
 import com.google.protobuf.Descriptors;
 import io.odpf.dagger.core.metrics.telemetry.TelemetryPublisher;
-import io.odpf.dagger.core.exception.DescriptorNotFoundException;
+import io.odpf.dagger.common.exceptions.DescriptorNotFoundException;
 import io.odpf.dagger.core.exception.InvalidConfigurationException;
 import io.odpf.dagger.common.metrics.managers.MeterStatsManager;
 import io.odpf.dagger.core.metrics.aspects.ExternalSourceAspects;
