+++ dagger-core/src/main/java/io/odpf/dagger/core/protohandler/TypeInformationFactory.java
@@ -1,6 +1,6 @@
 package io.odpf.dagger.core.protohandler;
 
-import io.odpf.dagger.core.exception.DescriptorNotFoundException;
+import io.odpf.dagger.common.exceptions.DescriptorNotFoundException;
 import com.google.protobuf.Descriptors;
 
 import org.apache.flink.api.common.typeinfo.TypeInformation;
