+++ dagger-core/src/test/java/io/odpf/dagger/core/processors/common/DescriptorManagerTest.java
@@ -2,7 +2,7 @@ package io.odpf.dagger.core.processors.common;
 
 import io.odpf.dagger.consumer.TestBookingLogMessage;
 import io.odpf.dagger.common.core.StencilClientOrchestrator;
-import io.odpf.dagger.core.exception.DescriptorNotFoundException;
+import io.odpf.dagger.common.exceptions.DescriptorNotFoundException;
 import com.gojek.de.stencil.StencilClientFactory;
 import com.gojek.de.stencil.client.StencilClient;
 import com.google.protobuf.Descriptors;
