+++ dagger-core/src/test/java/io/odpf/dagger/core/processors/external/http/HttpAsyncConnectorTest.java
@@ -3,7 +3,7 @@ package io.odpf.dagger.core.processors.external.http;
 import com.gojek.de.stencil.client.StencilClient;
 import io.odpf.dagger.consumer.TestBookingLogMessage;
 import io.odpf.dagger.common.core.StencilClientOrchestrator;
-import io.odpf.dagger.core.exception.DescriptorNotFoundException;
+import io.odpf.dagger.common.exceptions.DescriptorNotFoundException;
 import io.odpf.dagger.core.exception.InvalidConfigurationException;
 import io.odpf.dagger.core.exception.InvalidHttpVerbException;
 import io.odpf.dagger.common.metrics.managers.MeterStatsManager;
