+++ dagger-core/src/test/java/io/odpf/dagger/core/protohandler/TypeInformationFactoryTest.java
@@ -1,7 +1,7 @@
 package io.odpf.dagger.core.protohandler;
 
 import io.odpf.dagger.consumer.TestBookingLogKey;
-import io.odpf.dagger.core.exception.DescriptorNotFoundException;
+import io.odpf.dagger.common.exceptions.DescriptorNotFoundException;
 import com.google.protobuf.Descriptors;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.types.Row;
