+++ dagger-functions/src/main/java/io/odpf/dagger/functions/udfs/scalar/DartGet.java
@@ -30,7 +30,7 @@ public class DartGet extends ScalarUdf {
     public void open(FunctionContext context) throws Exception {
         super.open(context);
         MeterStatsManager meterStatsManager = new MeterStatsManager(context.getMetricGroup(), true);
-        meterStatsManager.register(UDF_TELEMETRY_GROUP_KEY, this.getClass().getSimpleName(), DartAspects.values());
+        meterStatsManager.register(UDF_TELEMETRY_GROUP_KEY, this.getName(), DartAspects.values());
         dataStore.setMeterStatsManager(meterStatsManager);
         dataStore.setGaugeStatsManager(getGaugeStatsManager());
     }
