+++ dagger-functions/src/main/java/io/odpf/dagger/functions/udfs/scalar/dart/store/gcs/GcsDataStore.java
@@ -79,7 +79,6 @@ public class GcsDataStore implements DataStore, Serializable {
             getMeterStatsManager().markEvent(DartAspects.DART_GCS_FETCH_FAILURES);
             e.printStackTrace();
         }
-
         return map;
     }
